+++ vmm-sys-util/src/windows/mod.rs
@@ -3,3 +3,4 @@
 pub mod epoll;
 pub mod event;
 pub mod eventfd;
+pub mod section;